backend/requirements.txt
# FastAPI & ASGI
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Redis
redis==5.0.1
aioredis==2.0.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic[email]==2.5.3
pydantic-settings==2.1.0

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# OpenAI / LLM
openai==1.10.0
langchain==0.1.4
langchain-openai==0.0.5

# Document Processing
pypdf==3.17.4
python-docx==1.1.0
python-pptx==0.6.23
pillow==10.2.0

# Storage
boto3==1.34.25
minio==7.2.3

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Development
black==24.1.0
isort==5.13.2
flake8==7.0.0
